Bug fix: memset of triangle repetition of the BBTree is fixed. Hack: In BDPT Proc, direct connection from light source to camera are disabled

+++ src/librender/skdtree.cpp
@@ -251,7 +251,7 @@ bool ShapeKDTree::ellipsoidParseKDTree(const KDNode* node, size_t& index, Ellips
 		if(m_BBTree->m_triangleRepetition[primIdx] <= 0){
 			SLog(EError, "Triangle repetition of %i th triangle is not properly measured", primIdx);
 		}
-		value = value*(u-l)*multiplier/m_BBTree->m_triangleRepetition[primIdx];
+		value = value*(u-l)*multiplier/(double)m_BBTree->m_triangleRepetition[primIdx];
 		return true;
 	}
 	e->cacheSetTriState(primIdx,Cache::EFails);
